web-server/requirements.txt
Flask_SocketIO==5.0.1
Flask==1.1.2
protobuf==3.15.8
google-cloud-storage==1.37.1
google-cloud-pubsub==2.4.1
gunicorn==20.1.0